원유현 exam2

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -8,15 +8,19 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.multipart.MultipartRequest;
 
 import com.example.demo.service.ArticleService;
 import com.example.demo.service.BoardService;
+import com.example.demo.service.MemberService;
 import com.example.demo.util.Ut;
 import com.example.demo.vo.Article;
 import com.example.demo.vo.Board;
 import com.example.demo.vo.Page;
 import com.example.demo.vo.ResultData;
 import com.example.demo.vo.Rq;
+
+import jakarta.servlet.http.HttpServletRequest;
 
 @Controller
 public class UsrArticleController {
@@ -29,6 +33,10 @@ public class UsrArticleController {
 
 	@Autowired
 	private BoardService boardService;
+	
+	@Autowired
+	private MemberService memberService;
+
 
 	// 액션 메소드
 	@RequestMapping("/usr/article/detail")
@@ -41,11 +49,30 @@ public class UsrArticleController {
 		}
 
 		// 게시글 db에서 가져오기 + 로그인 중인 아이디 권한체크까지 다 끝내고 가져온다.
-		Article article = articleService.getForPrintArticle(id);
+		Article article = articleService.getForPrintArticle(rq.getLoginedMemberId(), id);
 
 		model.addAttribute("article", article);
 
 		return "usr/article/detail";
+	}
+
+	@RequestMapping("/usr/article/doIncreaseHitCountRd")
+	@ResponseBody
+	public ResultData doIncreaseHitCountRd(int id) {
+
+		// 조회수 증가
+		ResultData increaseHitCountRd = articleService.increaseHitCount(id);
+
+		// 가져올 게시글 없는 경우 체크
+		if (increaseHitCountRd.isFail()) {
+			return increaseHitCountRd;
+		}
+
+		ResultData rd = ResultData.newData(increaseHitCountRd, "hitCount", articleService.getArticleHitCount(id));
+
+		rd.setData2("id", id);
+
+		return rd;
 	}
 
 	@RequestMapping("/usr/article/list")
@@ -100,12 +127,15 @@ public class UsrArticleController {
 	@RequestMapping("/usr/article/modify")
 	public String showModify(Integer id, Model model) {
 
+		// 로그인 정보 가져오기
+//		Rq rq = (Rq) req.getAttribute("rq");
+
 		// 게시글 존재여부 체크
 		if (id == null) {
 			return rq.historyBackOnView("없는 게시글이야");
 		}
 
-		Article article = articleService.getForPrintArticle(id);
+		Article article = articleService.getForPrintArticle(rq.getLoginedMemberId(), id);
 
 		model.addAttribute("article", article);
 
@@ -120,11 +150,18 @@ public class UsrArticleController {
 
 		Article article = articleService.getArticle(id); // 해당 게시글 가져오기
 
+		// 로그인 정보 가져오기
+//		Rq rq = (Rq) req.getAttribute("rq");
+
+		// 로그인 중인 아이디 권한체크(서비스에 요청)
+		ResultData loginedMemberCanModifyRd = articleService.userCanModify(rq.getLoginedMemberId(), article);
+
 		// 글 수정 작업
-
-		articleService.modifyArticle(id, title, body);
-
-		return Ut.jsReplace("s-1", "수정되었습니다", "../article/detail?id=" + id);
+		if (loginedMemberCanModifyRd.isSuccess()) {
+			articleService.modifyArticle(id, title, body);
+		}
+		return Ut.jsReplace(loginedMemberCanModifyRd.getResultCode(), loginedMemberCanModifyRd.getMsg(),
+				"../article/detail?id=" + id);
 	}
 
 	// 로그인 체크 -> 유무 체크 -> 권한 체크 -> 삭제
@@ -140,22 +177,32 @@ public class UsrArticleController {
 			return Ut.jsHistoryBack("F-1", Ut.f("%d번 글은 존재하지 않습니다", id));
 		}
 
+		// 로그인 정보 가져오기
+//		Rq rq = (Rq) req.getAttribute("rq");
+
+		// 로그인 중인 아이디 권한체크(서비스에 요청)
+		ResultData loginedMemberCanDeleteRd = articleService.userCanDelete(rq.getLoginedMemberId(), article);
+
 		// 글 삭제 작업
-
-		articleService.deleteArticle(id);
-
-		return Ut.jsReplace("S-2", "삭제되었습니다", "../article/list");
+		if (loginedMemberCanDeleteRd.isSuccess()) {
+			articleService.deleteArticle(id);
+		}
+		return Ut.jsReplace(loginedMemberCanDeleteRd.getResultCode(), loginedMemberCanDeleteRd.getMsg(),
+				"../article/list");
 	}
 
 	@RequestMapping("/usr/article/write")
-	public String showWrite() {
+	public String showJoin(Model model) {
 
+		int currentId = articleService.getCurrentArticleId();
+
+		model.addAttribute("currentId", currentId);
 		return "usr/article/write";
 	}
 
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
-	public String doWrite(String title, String body, int boardId) {
+	public String doWrite(HttpServletRequest req, int boardId, String title, String body) {
 		// 로그인 상태 체크 - 인터셉터에서
 
 		// 제목 내용 빈 칸 확인
@@ -170,10 +217,11 @@ public class UsrArticleController {
 //		Rq rq = (Rq) req.getAttribute("rq");
 
 		// 게시글 작성 작업
-		ResultData<Integer> writeArticleRd = articleService.writeArticle(title, body, boardId);
+		ResultData<Integer> writeArticleRd = articleService.writeArticle(title, body, rq.getLoginedMemberId(), boardId);
 
 		// 작성된 게시글 번호 가져오기
 		int id = (int) writeArticleRd.getData1();
+
 
 		return Ut.jsReplace(writeArticleRd.getResultCode(), writeArticleRd.getMsg(), "../article/detail?id=" + id);
 	}
